rework to use ConnectionHandler instead of ConnectionPool
- this provides compatibility with Rails 5.2 when using a forking server (like passenger, unicorn, and even puma)
