+++ lib/multidb.rb
@@ -7,5 +7,6 @@ require 'active_support/core_ext/module/aliasing'
 require_relative 'multidb/configuration'
 require_relative 'multidb/model_extensions'
 require_relative 'multidb/log_subscriber'
+require_relative 'multidb/candidate'
 require_relative 'multidb/balancer'
 require_relative 'multidb/version'
